Revert "Refactor user model to remove duplicate code"
