+++ src/main/java/bg/softuni/invoice/model/service/UserServiceModel.java
@@ -1,14 +1,53 @@
 package bg.softuni.invoice.model.service;
 
-import bg.softuni.invoice.model.base.UserBaseModel;
-
 import java.util.HashSet;
 import java.util.Set;
 
-public class UserServiceModel extends UserBaseModel {
+public class UserServiceModel {
 
+    private String id;
+    private String username;
+    private String firstName;
+    private String lastName;
     private String password;
-    private Set<RoleServiceModel> authorities = new HashSet<>();
+    private Set<RoleServiceModel> authorities;
+    private boolean enabled;
+
+    public UserServiceModel() {
+        setAuthorities(new HashSet<>());
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
 
     public String getPassword() {
         return password;
@@ -24,5 +63,13 @@ public class UserServiceModel extends UserBaseModel {
 
     public void setAuthorities(Set<RoleServiceModel> authorities) {
         this.authorities = authorities;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
     }
 }
