+++ src/main/java/bg/softuni/invoice/model/view/UserViewModel.java
@@ -1,13 +1,47 @@
 package bg.softuni.invoice.model.view;
 
-import bg.softuni.invoice.model.base.UserBaseModel;
-
-import java.util.HashSet;
 import java.util.Set;
 
-public class UserViewModel extends UserBaseModel {
+public class UserViewModel {
 
-    private Set<String> authorities = new HashSet<>();
+    private String id;
+    private String username;
+    private String firstName;
+    private String lastName;
+    private Set<String> authorities;
+    private boolean enabled;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
 
     public Set<String> getAuthorities() {
         return authorities;
@@ -15,5 +49,13 @@ public class UserViewModel extends UserBaseModel {
 
     public void setAuthorities(Set<String> authorities) {
         this.authorities = authorities;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
     }
 }
