+++ src/test/java/bg/softuni/invoice/model/view/UserViewModelTest.java
@@ -7,7 +7,6 @@ import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -153,13 +152,12 @@ class UserViewModelTest {
     }
 
     @Test
-    void authorities_whenNotSet_shouldReturnEmptyCollection() {
+    void authorities_whenNotSet_shouldReturnNull() {
         UserViewModel userViewModel = new UserViewModel();
 
         Set<String> actualAuthorities = userViewModel.getAuthorities();
 
-        assertNotNull(actualAuthorities);
-        assertTrue(actualAuthorities.isEmpty());
+        assertNull(actualAuthorities);
     }
 
     @Test
